# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Les Casaniers API",
        "description": "Documentation de l'API pour le projet Les Casaniers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Principal"
        }
    ],
    "paths": {
...
        "/api/produits/{produitId}/attributes/sync": {
            "post": {
                "tags": [
                    "Attributs Produits"
                ],
                "summary": "Synchroniser les attributs d'un produit (Admin)",
                "description": "Remplace tous les attributs existants par la nouvelle liste fournie.",
                "operationId": "8e30dc73291b74bcf9d65732fef5a2aa",
                "parameters": [
                    {
                        "name": "produitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "cle_attr": {
                                            "type": "string",
                                            "example": "socket"
                                        },
                                        "valeur_attr": {
                                            "type": "string",
                                            "example": "LGA1700"
                                        },
                                        "libelle_attr": {
                                            "type": "string",
                                            "example": "Socket CPU"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/attributes/standard-keys": {
            "get": {
                "tags": [
                    "Attributs Produits"
                ],
                "summary": "Récupérer le dictionnaire des clés standard (Admin)",
                "operationId": "33e2b17d588255bb15d146c618b366ae",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Liste des catégories racines",
                "operationId": "b77b1daca69d91daf0fb0560ae4d961f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une catégorie (Admin)",
                "operationId": "336078eeb5da88674a00f6e8666ec9c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "type"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/{slug}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Détails d'une catégorie par slug",
                "operationId": "227717dda19dd9a50517424bccfb920b",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie (Admin)",
                "operationId": "cf935727a14dda7603614c02178a7a38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie (Admin)",
                "operationId": "fb09c985ece1b538e0a0b16b5a77fe71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/reorder": {
            "patch": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Réorganiser les catégories (Admin)",
                "operationId": "faac2cdef4134f01d01badec67a067d8",
                "responses": {
                    "200": {
                        "description": "Réorganisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/produits/{produitId}/images": {
            "post": {
                "tags": [
                    "Images Produits"
                ],
                "summary": "Uploader une image pour un produit (Admin)",
                "operationId": "2d691ecea4242de6eab5b975c190e469",
                "parameters": [
                    {
                        "name": "produitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt": {
                                        "type": "string"
                                    },
                                    "ordre": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Upload réussi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/images/{id}": {
            "delete": {
                "tags": [
                    "Images Produits"
                ],
                "summary": "Supprimer une image (Admin)",
                "operationId": "415ef54e96da0857d1acfb7eba08948d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/produits/{produitId}/images/{imageId}/set-main": {
            "patch": {
                "tags": [
                    "Images Produits"
                ],
                "summary": "Définir une image comme principale (Admin)",
                "operationId": "e0df1e5866fdfeafbf1a877195eecd80",
                "parameters": [
                    {
                        "name": "produitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "imageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}